fix warnings/gcc build

+++ src/audio.hh
@@ -25,7 +25,7 @@ struct MixerInterface
     void (*play)(Mixer*, String);
     void (*pause)(Mixer* s, bool bPause);
     void (*togglePause)(Mixer* s);
-    void (*changeSampleRate)(Mixer* s, int sampleRate, bool bSave);
+    void (*changeSampleRate)(Mixer* s, u64 sampleRate, bool bSave);
     void (*seekMS)(Mixer* s, u64 ms);
     void (*seekLeftMS)(Mixer* s, u64 ms);
     void (*seekRightMS)(Mixer* s, u64 ms);
@@ -54,7 +54,7 @@ ADT_NO_UB constexpr void MixerDestroy(Mixer* s) { s->pVTable->destroy(s); }
 ADT_NO_UB constexpr void MixerPlay(Mixer* s, String sPath) { s->pVTable->play(s, sPath); }
 ADT_NO_UB constexpr void MixerPause(Mixer* s, bool bPause) { s->pVTable->pause(s, bPause); }
 ADT_NO_UB constexpr void MixerTogglePause(Mixer* s) { s->pVTable->togglePause(s); }
-ADT_NO_UB constexpr void MixerChangeSampleRate(Mixer* s, int sampleRate, bool bSave) { s->pVTable->changeSampleRate(s, sampleRate, bSave); }
+ADT_NO_UB constexpr void MixerChangeSampleRate(Mixer* s, u64 sampleRate, bool bSave) { s->pVTable->changeSampleRate(s, sampleRate, bSave); }
 ADT_NO_UB constexpr void MixerSeekMS(Mixer* s, u64 ms) { s->pVTable->seekMS(s, ms); }
 ADT_NO_UB constexpr void MixerSeekLeftMS(Mixer* s, u64 ms) { s->pVTable->seekLeftMS(s, ms); }
 ADT_NO_UB constexpr void MixerSeekRightMS(Mixer* s, u64 ms) { s->pVTable->seekRightMS(s, ms); }
@@ -87,7 +87,7 @@ constexpr void DummyMixerDestroy([[maybe_unused]] DummyMixer* s) {}
 constexpr void DummyMixerPlay([[maybe_unused]] DummyMixer* s, [[maybe_unused]] String sPath) {}
 constexpr void DummyMixerPause([[maybe_unused]] DummyMixer* s, [[maybe_unused]] bool bPause) {}
 constexpr void DummyMixerTogglePause([[maybe_unused]] DummyMixer* s) {}
-constexpr void DummyMixerChangeSampleRate([[maybe_unused]] DummyMixer* s, [[maybe_unused]] int sampleRate, [[maybe_unused]] bool bSave) {}
+constexpr void DummyMixerChangeSampleRate([[maybe_unused]] DummyMixer* s, [[maybe_unused]] u64 sampleRate, [[maybe_unused]] bool bSave) {}
 constexpr long DummyMixerGetCurrentMS([[maybe_unused]] Mixer* s) { return {}; }
 constexpr long DummyMixerGetMaxMS([[maybe_unused]] Mixer* s) { return {}; }
 constexpr void DummyMixerSeekMS([[maybe_unused]] DummyMixer* s, [[maybe_unused]] u64 ms) {}
